Change location for icons + added solidSvg plugin
