+++ app.config.ts
@@ -1,23 +1,22 @@
 import { defineConfig } from '@solidjs/start/config';
-import { configDefaults } from 'vitest/config'
+import { configDefaults } from 'vitest/config';
 import autoprefixer from 'autoprefixer';
+import solidSvg from 'vite-plugin-solid-svg';
 
 export default defineConfig({
   server: {
     preset: 'aws-lambda',
-    inlineDynamicImports: true
+    inlineDynamicImports: true,
   },
   vite: {
+    plugins: [solidSvg()],
     css: {
       postcss: {
         plugins: [autoprefixer({})],
       },
     },
     test: {
-      exclude:[
-        ...configDefaults.exclude, 
-        'cdk/*'
-      ],
+      exclude: [...configDefaults.exclude, 'cdk/*'],
       environment: 'jsdom',
       globals: true,
       setupFiles: ['node_modules/@testing-library/jest-dom/vitest'],
