+++ package.json
@@ -35,6 +35,7 @@
     "ts-jest": "^29.1.1",
     "typescript": "^5.7.2",
     "typescript-eslint": "^8.16.0",
+    "vite-plugin-solid-svg": "^0.8.1",
     "vitest": "^1.2.1"
   },
   "dependencies": {
