+++ src/components/DetailOverview/index.tsx
@@ -16,7 +16,7 @@ import { SensorType } from './SensorType';
 import '~/assets/scss/components/detail-overview.scss';
 import { DetailOverviewDefinition } from './types';
 import { sensorNodeLists } from '~/db/lists';
-import { LocationLicenses } from '../LocationLicenses';
+import { License } from '../License';
 
 interface ListsDefinition {
   sensorNodesId: number;
@@ -70,15 +70,12 @@ function LocationListsFallback() {
 
 export function DetailOverview(props: DetailOverviewDefinition) {
   const pageLocation = useLocation();
-  
 
   return (
     <section class="detail-overview">
       <div class="detail-overview__title">
         <div>
-          <span class="type-subtitle-3 text-smoke-120">
-            {props?.name}
-          </span>
+          <span class="type-subtitle-3 text-smoke-120">{props?.name}</span>
           <h1 class="type-display-1 text-sky-120">
             {props.name || 'No label'}
           </h1>
@@ -150,7 +147,7 @@ export function DetailOverview(props: DetailOverviewDefinition) {
                 <td>Provider</td>
                 <td>{props.provider?.name}</td>
               </tr>
-              <LocationLicenses {...props} />
+              <License {...props} />
             </tbody>
           </table>
         </div>
